dom/tests/mochitest/mouselock/test_jbraffoul.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=92264
-->
<head>
  <title>Test for Bug 633602</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
   <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">Mozilla Bug 633602</a>
   <p id="display"></p>
   <div id="content" style="display: none">
   </div>
	
   <canvas id="canvas" width="150" height="150"></canvas>

   <pre id="test">
      <script>
         SimpleTest.waitForExplicitFinish();
         var canvas = document.getElementById("canvas");
         var pointer = navigator.pointer;
         var mouseMoved = false;
         SpecialPowers.setBoolPref("full-screen-api.allow-trusted-requests-only", false);

         SimpleTest.waitForFocus(function() {
            
            // If mouse moves while mouse is locked - print screen coords
            canvas.addEventListener('mousemove', function(e) {
               if(mouseMoved == false) {
                  var clientX1 = e.clientX;
                  var clientY1 = e.clientY;
                  mouseMoved = true;
               } else {
                  var clientX2 = e.clientX;
                  var clientY2 = e.clientY;
               }
               
            }, false);
         
            // When fullscreen is enabled on canvas - lock the mouse 
            document.addEventListener('mozfullscreenchange', function() {
               if(document.mozFullScreen) {
                  if(document.mozFullScreenElement == canvas) {
                     pointer.lock(canvas);
                  }
               }
            }, false);
            
            canvas.mozRequestFullScreen();
            synthesizeMouse(canvas, 1000, 1000, {});
            document.mozCancelFullScreen();
            
            is(clientX1, clientX2, "error X");
            is(clientY1, clientY2, "error Y");
           
            is(1,1,"1 == 1");
         }
         SimpleTest.finish();
      </script>
   </pre>
</body>
</html>
